Use gradlew for hugo-example (not gradle of OS)
There is a bug when the following tasks are executed:

cleanExample
assembleExample
installExample
In these tasks the gradle of the system is used instead of gradlew.
This leads to strange errors when the two versions are different.

The simple correction consisted on changing the command to execute
from 'gradle to '../gradlew'.

task cleanExample(type: Exec) {
    executable = '../gradlew'
    workingDir = project.file('hugo-example')
    ...